pyproject.toml
[tool.poetry]
name = "nexusbot"
version = "0.1.0"
description = ""
authors = ["ryanp <ryanpey>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
